backend/requirements.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
django-environ==0.11.*
Pillow==10.3.*
